+++ CHANGELOG.md
@@ -9,7 +9,6 @@
 <!-- Fixme mpv setting up procedure, such as write test scripts for correct pattern line pattern (must has 1 capture group -->
 <!-- Todo allow set PASS message as an arg, otherwise checking same file for different purpose print confusing message -->
 <!-- todo: for commit message, distinguish Major/Minor/Patch Release -->
-<!-- Bug: when matching for AMs, make sure space around keywords -->
 <!-- todo UI to interactively setting env vars -->
 
 
@@ -33,7 +32,12 @@
 
 ### Deprecated
 ### Removed
+
+- branch skip feature
+
 ### Fixed
+
+- BP pattern to not match unintentional pattern